+++ packages/core-mobile/app/new/routes/(signedIn)/_layout.tsx
@@ -7,17 +7,16 @@ import {
   stackNavigatorScreenOptions,
   stackScreensOptions
 } from 'common/consts/screenOptions'
-
-import { usePerformAfterLoginFlows } from 'common/hooks/usePerformAfterLoginFlows'
 import { BridgeProvider } from 'features/bridge/contexts/BridgeContext'
 import { CollectiblesProvider } from 'features/portfolio/collectibles/CollectiblesContext'
 import { MigrateFavoriteIds } from 'new/common/components/MigrateFavoriteIds'
 import { NavigationPresentationMode } from 'new/common/types'
 import React, { useEffect } from 'react'
-import { useSelector } from 'react-redux'
+import { useDispatch, useSelector } from 'react-redux'
 import { selectActiveAccount } from 'store/account'
 import { selectWalletState } from 'store/app'
 import { WalletState } from 'store/app/types'
+import { afterLoginFlowsRequested } from 'store/notifications'
 
 const PolyfillCrypto = React.lazy(() => import('react-native-webview-crypto'))
 
@@ -28,6 +27,7 @@ export const unstable_settings = {
 export default function WalletLayout(): JSX.Element {
   const walletState = useSelector(selectWalletState)
   const activeAccount = useSelector(selectActiveAccount)
+  const dispatch = useDispatch()
 
   usePreventRemove(walletState === WalletState.ACTIVE, () => {
     // TODO: uncomment this after we fix the multiple back() calls
@@ -36,13 +36,11 @@ export default function WalletLayout(): JSX.Element {
     // BackHandler.exitApp()
   })
 
-  const performAfterLoginFlows = usePerformAfterLoginFlows()
-
   useEffect(() => {
     if (walletState !== WalletState.ACTIVE || !activeAccount) return
 
-    performAfterLoginFlows()
-  }, [walletState, activeAccount, performAfterLoginFlows])
+    dispatch(afterLoginFlowsRequested)
+  }, [walletState, activeAccount, dispatch])
 
   return (
     <BridgeProvider>
